components: refactor code
