--- src/components/Signup/styles.module.scss
@@ -1,93 +1,0 @@
-@import 'theme/colors';
-@import 'theme/fonts';
-
-.signUpContainer {
-	align-items: center;
-	background-color: color(background);
-	display: flex;
-	flex-shrink: 0;
-	flex-direction: column;
-	font-family: font(primary);
-	justify-content: center;
-	min-height: 100%;
-	padding: 2rem 10rem;
-	width: 100%;
-
-	@media (max-width: 1600px) {
-		padding: 2rem 4rem;
-	}
-	@media (max-width: 800px) {
-		padding: 2rem 6rem;
-	}
-
-	@media (max-width: 600px) {
-		padding: 2rem 4rem;
-	}
-
-	@media (max-width: 520px) {
-		padding: 2rem 2rem;
-	}
-
-	.logo {
-		margin: 0 auto 4rem auto;
-	}
-
-	.heading {
-		text-align: center;
-		& > h1 {
-			color: color(white);
-			font-size: size(h2);
-			font-weight: weight(medium);
-			margin: 0;
-		}
-		& > p {
-			color: color(textLight);
-			font-size: size(medium);
-			font-weight: weight(medium);
-		}
-	}
-	.inputContainer {
-		align-items: flex-end;
-		display: flex;
-		flex-direction: column;
-		margin: 0.5rem 0 4rem 0;
-		width: 100%;
-	}
-
-	.responseError {
-		text-align: center;
-
-		&.active {
-			margin: 1rem 0;
-		}
-	}
-
-	.bottomContainer {
-		width: 100%;
-
-		.google {
-			align-items: center;
-			background-color: color(white);
-			color: color(textDark);
-			display: flex;
-			justify-content: center;
-
-			.googleLogo {
-				height: 1.5rem;
-				margin-right: 0.5rem;
-			}
-		}
-
-		.noAccount {
-			color: color(white);
-			font-size: size(medium);
-			font-weight: weight(regular);
-			margin-top: 2rem;
-			text-align: center;
-			& > a {
-				color: color(accent);
-				margin-left: 0.1rem;
-			}
-		}
-	}
-}
