+++ src/components/auth/GoogleSignInButton/styles.module.scss
@@ -1,0 +1,14 @@
+@import 'theme/colors';
+
+.google {
+	align-items: center;
+	background-color: color(white);
+	color: color(textDark);
+	display: flex;
+	justify-content: center;
+
+	.googleLogo {
+		height: 1.5rem;
+		margin-right: 0.5rem;
+	}
+}
