+++ src/components/auth/LoginBrandInfo/styles.module.scss
@@ -4,7 +4,7 @@
 .infoContainer {
 	align-items: center;
 	background-blend-mode: overlay;
-	background-image: url('../../assets/images/login-bg.jpg'),
+	background-image: url('../../../assets/images/login-bg.jpg'),
 		linear-gradient(rgba(color(black), 0.4), rgba(color(black), 1));
 	background-size: cover;
 	color: color(white);
