+++ src/components/auth/Signup/styles.module.scss
@@ -1,0 +1,5 @@
+@import '../styles/auth.scss';
+
+.signUpContainer {
+	@extend .container;
+}
